refactor(mcp): split runtime.rs code-analysis helpers into code_analysis.rs (readability ratchet)
