+++ crates/temper-mcp/src/runtime.rs
@@ -14,6 +14,7 @@ use tokio::io::{self, AsyncBufRead, AsyncWrite, AsyncWriteExt, BufReader};
 use tokio::sync::mpsc;
 
 use super::McpConfig;
+use super::code_analysis::{extract_temper_call_metadata, extract_trajectory_actions_from_code};
 use super::protocol::{dispatch_json_value, json_rpc_error};
 use crate::elicit::{
     ClientRequester, DeniedDecision, PendingClientRequests, denial_from_dispatch_value,
@@ -586,369 +587,6 @@ impl RuntimeContext {
 struct TrajectoryUploadError {
     message: String,
     retryable: bool,
-}
-
-fn extract_trajectory_actions_from_code(code: &str) -> Vec<Value> {
-    let mut actions = Vec::new();
-    let mut cursor = 0usize;
-    let needle = "temper.action";
-
-    while let Some(found) = code[cursor..].find(needle) {
-        let method_start = cursor + found + needle.len();
-        let mut open = method_start;
-        while open < code.len()
-            && code
-                .as_bytes()
-                .get(open)
-                .is_some_and(|b| b.is_ascii_whitespace())
-        {
-            open += 1;
-        }
-        if code.as_bytes().get(open) != Some(&b'(') {
-            cursor = method_start;
-            continue;
-        }
-
-        let Some(close) = find_matching_paren(code, open) else {
-            break;
-        };
-
-        let args = split_top_level_args(&code[open + 1..close]);
-        let (action_idx, params_idx) =
-            if args.len() >= 5 && parse_python_string_literal(args[3]).is_some() {
-                (3usize, 4usize)
-            } else {
-                (2usize, 3usize)
-            };
-
-        if args.len() > action_idx
-            && let Some(action_name) = parse_python_string_literal(args[action_idx])
-        {
-            let params = args
-                .get(params_idx)
-                .and_then(|raw| parse_python_json_value(raw))
-                .unwrap_or_else(|| serde_json::json!({}));
-            actions.push(serde_json::json!({
-                "action": action_name,
-                "params": params,
-            }));
-        }
-
-        cursor = close + 1;
-    }
-
-    actions
-}
-
-#[derive(Debug, Clone, Default)]
-struct TemperCallMetadata {
-    /// Tenant referenced by the call, when the call uses the tenant-first
-    /// signature. Used only as a cross-tenant observability signal — never to
-    /// route trajectory storage (ARN-222).
-    tenant: Option<String>,
-}
-
-fn extract_temper_call_metadata(code: &str) -> Vec<TemperCallMetadata> {
-    let mut out = Vec::new();
-    out.extend(extract_temper_action_metadata(code));
-    out.extend(extract_temper_create_metadata(code));
-    out
-}
-
-fn extract_temper_action_metadata(code: &str) -> Vec<TemperCallMetadata> {
-    extract_call_metadata(code, "temper.action", |args| {
-        // New signature: temper.action(tenant, entity_type, id, action, params).
-        // Only the tenant is retained; the legacy signature carries no tenant.
-        let tenant = (args.len() >= 5)
-            .then(|| parse_python_string_literal(args[0]))
-            .flatten();
-        TemperCallMetadata { tenant }
-    })
-}
-
-fn extract_temper_create_metadata(code: &str) -> Vec<TemperCallMetadata> {
-    extract_call_metadata(code, "temper.create", |args| {
-        // New signature: temper.create(tenant, entity_type, fields). Only the
-        // tenant is retained; the legacy signature carries no tenant.
-        let tenant = (args.len() >= 3)
-            .then(|| parse_python_string_literal(args[0]))
-            .flatten();
-        TemperCallMetadata { tenant }
-    })
-}
-
-fn extract_call_metadata<F>(code: &str, needle: &str, mapper: F) -> Vec<TemperCallMetadata>
-where
-    F: Fn(Vec<&str>) -> TemperCallMetadata,
-{
-    let mut out = Vec::new();
-    let mut cursor = 0usize;
-
-    while let Some(found) = code[cursor..].find(needle) {
-        let method_start = cursor + found + needle.len();
-        let mut open = method_start;
-        while open < code.len()
-            && code
-                .as_bytes()
-                .get(open)
-                .is_some_and(|b| b.is_ascii_whitespace())
-        {
-            open += 1;
-        }
-        if code.as_bytes().get(open) != Some(&b'(') {
-            cursor = method_start;
-            continue;
-        }
-
-        let Some(close) = find_matching_paren(code, open) else {
-            break;
-        };
-        let args = split_top_level_args(&code[open + 1..close]);
-        out.push(mapper(args));
-        cursor = close + 1;
-    }
-
-    out
-}
-
-fn find_matching_paren(input: &str, open_idx: usize) -> Option<usize> {
-    let mut depth = 0i32;
-    let mut in_quote: Option<char> = None;
-    let mut escaped = false;
-
-    for (offset, ch) in input[open_idx..].char_indices() {
-        let idx = open_idx + offset;
-        if let Some(quote) = in_quote {
-            if escaped {
-                escaped = false;
-                continue;
-            }
-            if ch == '\\' {
-                escaped = true;
-                continue;
-            }
-            if ch == quote {
-                in_quote = None;
-            }
-            continue;
-        }
-
-        match ch {
-            '\'' | '"' => in_quote = Some(ch),
-            '(' => depth += 1,
-            ')' => {
-                depth -= 1;
-                if depth == 0 {
-                    return Some(idx);
-                }
-            }
-            _ => {}
-        }
-    }
-
-    None
-}
-
-fn split_top_level_args(input: &str) -> Vec<&str> {
-    let mut parts = Vec::new();
-    let mut start = 0usize;
-    let mut depth_paren = 0i32;
-    let mut depth_brace = 0i32;
-    let mut depth_bracket = 0i32;
-    let mut in_quote: Option<char> = None;
-    let mut escaped = false;
-
-    for (idx, ch) in input.char_indices() {
-        if let Some(quote) = in_quote {
-            if escaped {
-                escaped = false;
-                continue;
-            }
-            if ch == '\\' {
-                escaped = true;
-                continue;
-            }
-            if ch == quote {
-                in_quote = None;
-            }
-            continue;
-        }
-
-        match ch {
-            '\'' | '"' => in_quote = Some(ch),
-            '(' => depth_paren += 1,
-            ')' => depth_paren -= 1,
-            '{' => depth_brace += 1,
-            '}' => depth_brace -= 1,
-            '[' => depth_bracket += 1,
-            ']' => depth_bracket -= 1,
-            ',' if depth_paren == 0 && depth_brace == 0 && depth_bracket == 0 => {
-                parts.push(input[start..idx].trim());
-                start = idx + 1;
-            }
-            _ => {}
-        }
-    }
-
-    if start <= input.len() {
-        let tail = input[start..].trim();
-        if !tail.is_empty() {
-            parts.push(tail);
-        }
-    }
-    parts
-}
-
-fn parse_python_string_literal(raw: &str) -> Option<String> {
-    let s = raw.trim();
-    if s.len() < 2 {
-        return None;
-    }
-    let quote = s.chars().next()?;
-    if (quote != '\'' && quote != '"') || !s.ends_with(quote) {
-        return None;
-    }
-
-    let mut out = String::new();
-    let mut escaped = false;
-    for ch in s[1..s.len() - 1].chars() {
-        if escaped {
-            let mapped = match ch {
-                'n' => '\n',
-                'r' => '\r',
-                't' => '\t',
-                '\\' => '\\',
-                '\'' => '\'',
-                '"' => '"',
-                other => other,
-            };
-            out.push(mapped);
-            escaped = false;
-            continue;
-        }
-        if ch == '\\' {
-            escaped = true;
-            continue;
-        }
-        out.push(ch);
-    }
-    if escaped {
-        out.push('\\');
-    }
-    Some(out)
-}
-
-fn parse_python_json_value(raw: &str) -> Option<Value> {
-    let trimmed = raw.trim();
-    if trimmed.is_empty() {
-        return Some(serde_json::json!({}));
-    }
-    if let Ok(v) = serde_json::from_str::<Value>(trimmed) {
-        return Some(v);
-    }
-    let normalized = normalize_pythonish_json(trimmed);
-    serde_json::from_str::<Value>(&normalized).ok()
-}
-
-fn normalize_pythonish_json(input: &str) -> String {
-    let mut quoted = String::with_capacity(input.len());
-    let mut in_single = false;
-    let mut in_double = false;
-    let mut escaped = false;
-
-    for ch in input.chars() {
-        if in_single {
-            if escaped {
-                quoted.push(ch);
-                escaped = false;
-                continue;
-            }
-            match ch {
-                '\\' => escaped = true,
-                '\'' => {
-                    in_single = false;
-                    quoted.push('"');
-                }
-                '"' => quoted.push_str("\\\""),
-                _ => quoted.push(ch),
-            }
-            continue;
-        }
-
-        if in_double {
-            quoted.push(ch);
-            if escaped {
-                escaped = false;
-            } else if ch == '\\' {
-                escaped = true;
-            } else if ch == '"' {
-                in_double = false;
-            }
-            continue;
-        }
-
-        match ch {
-            '\'' => {
-                in_single = true;
-                quoted.push('"');
-            }
-            '"' => {
-                in_double = true;
-                quoted.push('"');
-            }
-            _ => quoted.push(ch),
-        }
-    }
-
-    let mut out = String::with_capacity(quoted.len());
-    let mut token = String::new();
-    let mut in_string = false;
-    let mut esc = false;
-
-    let flush_token = |token: &mut String, out: &mut String| {
-        if token.is_empty() {
-            return;
-        }
-        match token.as_str() {
-            "True" => out.push_str("true"),
-            "False" => out.push_str("false"),
-            "None" => out.push_str("null"),
-            _ => out.push_str(token),
-        }
-        token.clear();
-    };
-
-    for ch in quoted.chars() {
-        if in_string {
-            out.push(ch);
-            if esc {
-                esc = false;
-            } else if ch == '\\' {
-                esc = true;
-            } else if ch == '"' {
-                in_string = false;
-            }
-            continue;
-        }
-
-        if ch == '"' {
-            flush_token(&mut token, &mut out);
-            in_string = true;
-            out.push(ch);
-            continue;
-        }
-
-        if ch.is_ascii_alphanumeric() || ch == '_' {
-            token.push(ch);
-            continue;
-        }
-
-        flush_token(&mut token, &mut out);
-        out.push(ch);
-    }
-    flush_token(&mut token, &mut out);
-
-    out
 }
 
 /// Run the MCP server on stdio with JSON-RPC over newline-delimited JSON.
